// frontend/src/app/shared/enums/elastic-operators.enum.ts
export enum ElasticOperatorsEnum {
  IS = 'IS',
  IS_NOT = 'IS_NOT',
  IS_ONE_OF = 'IS_ONE_OF',
  IS_NOT_ONE_OF = 'IS_NOT_ONE_OF',
  EXIST = 'EXIST',
  DOES_NOT_EXIST = 'DOES_NOT_EXIST',
  IS_BETWEEN = 'IS_BETWEEN',
  IS_NOT_BETWEEN = 'IS_NOT_BETWEEN',
  CONTAIN = 'CONTAIN',
  DOES_NOT_CONTAIN = 'DOES_NOT_CONTAIN',
  IS_IN_FIELD = 'IS_IN_FIELDS',
  IS_NOT_FIELD = 'IS_NOT_IN_FIELDS',
  ENDS_WITH = 'ENDS_WITH',
  NOT_ENDS_WITH = 'NOT_ENDS_WITH',
  START_WITH = 'START_WITH',
  NOT_START_WITH = 'NOT_START_WITH',
  IS_ONE_OF_TERMS = 'IS_ONE_OF_TERMS'
}